requirements.txt
fastapi
uvicorn
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
newspaper3k
python-dotenv
lxml[html_clean]